tests: inbox: reorganize tests
